requirements.txt
Pillow>=4.0.0,<5.0.0
numpy>=1.11.0,<1.12.0
